abstract out refreshPreview
